+++ docs/prime/factor.py
@@ -1,9 +1,9 @@
+from __future__ import annotations
 import json
 from pathlib import Path
-from typing import List
 
 
-def factor(x: int) -> List[int]:
+def factor(x: int) -> list[int]:
     for f in range(2, x // 2 + 1):
         if x % f == 0:
             return [f] + factor(x // f)