+++ myqueue/task.py
@@ -55,7 +55,7 @@ class Task:
                  creates: list[str],
                  notifications: str = '',
                  state: State = State.undefined,
-                 id: int = 0,
+                 id: str = '0',
                  error: str = '',
                  memory_usage: int = 0,
                  tqueued: float = 0.0,
@@ -97,6 +97,10 @@ class Task:
     def name(self) -> str:
         return f'{self.cmd.name}.{self.id}'
 
+    @ property
+    def int_id(self) -> int:
+        return int(self.id.split('.')[0])
+
     def running_time(self, t: float = None) -> float:
         if self.state in ['CANCELED', 'queued', 'hold']:
             dt = 0.0
@@ -124,7 +128,7 @@ class Task:
         if self.notifications:
             info.append(self.notifications)
 
-        return [str(self.id),
+        return [self.id,
                 str(self.folder) + '/',
                 self.cmd.short_name,
                 ' '.join(self.cmd.args),
@@ -240,7 +244,7 @@ class Task:
                     creates.split(','),
                     notifications,
                     State[state],
-                    int(id),
+                    id,
                     error,
                     memory_usage,
                     tqueued, trunning, tstop)
@@ -270,12 +274,15 @@ class Task:
             folder = root / f
             deps = [root / dep for dep in dct.pop('deps')]
 
+        id = str(dct.pop('id'))
+
         return Task(cmd=create_command(**dct.pop('cmd')),
                     resources=Resources(**dct.pop('resources')),
                     state=State[dct.pop('state')],
                     folder=folder,
                     deps=deps,
                     notifications=dct.pop('notifications', ''),
+                    id=id,
                     **dct)
 
     def infolder(self, folder: Path, recursive: bool) -> bool:
@@ -352,7 +359,7 @@ class Task:
             self.error = lines[-1]
         return False
 
-    def ideps(self, map: dict[Path, 'Task']) -> Iterator['Task']:
+    def ideps(self, map: dict[Path, Task]) -> Iterator[Task]:
         """Yield task and its dependencies."""
         yield self
         for dname in self.deps:
@@ -376,7 +383,7 @@ class Task:
             queue.submit([self])
 
     def find_dependents(self,
-                        tasks: Sequence['Task']) -> Iterator['Task']:
+                        tasks: Sequence[Task]) -> Iterator[Task]:
         """Yield dependents."""
         for task in tasks:
             if self.dname in task.deps and self is not task:
@@ -384,7 +391,7 @@ class Task:
                 yield from task.find_dependents(tasks)
 
     def cancel_dependents(self,
-                          tasks: Sequence['Task'],
+                          tasks: Sequence[Task],
                           t: float = 0.0) -> int:
         """Cancel dependents."""
         ncancel = 0
@@ -396,6 +403,12 @@ class Task:
 
     def run(self) -> None:
         self.result = self.cmd.run()
+
+    def get_venv_activation_line(self) -> str:
+        if self.activation_script:
+            return (f'source {self.activation_script}\n'
+                    f'echo "venv: {self.activation_script}"\n')
+        return ''
 
 
 def task(cmd: str,